directory needs to be created first, in some cases
for new installations, it's necessary to create the Bundles directory.

I'm submitting this pull request with the assumption that the last one was accepted.  If you want to reject the last one, I can change this to use Avian (or write something more complicated that handles whatever situation requires it).